+++ client/app/reader/DecisionReviewer.jsx
@@ -187,7 +187,7 @@ DecisionReviewer.propTypes = {
 
 const mapStateToProps = (state) => {
   return {
-    documentFilters: state.readerReducer.ui.pdfList.filters,
+    documentFilters: state.documentList.pdfList.filters,
     storeDocuments: state.documents,
     pdf: state.readerReducer.ui.pdf,
     isPlacingAnnotation: state.annotationLayer.isPlacingAnnotation
