+++ client/app/reader/DocumentList/actionTypes.js
@@ -1,0 +1,37 @@
+export const SET_SORT = 'SET_SORT';
+export const CLEAR_CATEGORY_FILTER = 'CLEAR_CATEGORY_FILTER';
+export const SET_CATEGORY_FILTER = 'SET_CATEGORY_FILTER';
+export const TOGGLE_FILTER_DROPDOWN = 'TOGGLE_FILTER_DROPDOWN';
+export const SET_TAG_FILTER = 'SET_TAG_FILTER';
+export const CLEAR_TAG_FILTER = 'CLEAR_TAG_FILTER';
+export const SET_DOC_LIST_SCROLL_POSITION = 'SET_DOC_LIST_SCROLL_POSITION';
+export const SET_VIEWING_DOCUMENTS_OR_COMMENTS = 'SET_VIEWING_DOCUMENTS_OR_COMMENTS';
+export const SET_SEARCH = 'SET_SEARCH';
+export const CLEAR_ALL_SEARCH = 'CLEAR_ALL_SEARCH';
+export const CLEAR_ALL_FILTERS = 'CLEAR_ALL_FILTERS';
+export const RECEIVE_MANIFESTS = 'RECEIVE_MANIFESTS';
+export const UPDATE_FILTERED_RESULTS = 'UPDATE_FILTERED_RESULTS';
+
+export const DOCUMENTS_OR_COMMENTS_ENUM = {
+  DOCUMENTS: 'documents',
+  COMMENTS: 'comments'
+};
+
+export const ASSIGN_DOCUMENTS = 'ASSIGN_DOCUMENTS';
+export const RECEIVE_DOCUMENTS = 'RECEIVE_DOCUMENTS';
+export const TOGGLE_DOCUMENT_CATEGORY_FAIL = 'TOGGLE_DOCUMENT_CATEGORY_FAIL';
+export const TOGGLE_DOCUMENT_CATEGORY = 'TOGGLE_DOCUMENT_CATEGORY';
+export const TOGGLE_COMMENT_LIST = 'TOGGLE_COMMENT_LIST';
+export const ROTATE_PDF_DOCUMENT = 'ROTATE_PDF_DOCUMENT';
+
+export const ROTATION_INCREMENTS = 90;
+
+export const SELECT_CURRENT_VIEWER_PDF = 'SELECT_CURRENT_VIEWER_PDF';
+
+export const REQUEST_NEW_TAG_CREATION_SUCCESS = 'REQUEST_NEW_TAG_CREATION_SUCCESS';
+export const REQUEST_NEW_TAG_CREATION_FAILURE = 'REQUEST_NEW_TAG_CREATION_FAILURE';
+export const REQUEST_NEW_TAG_CREATION = 'REQUEST_NEW_TAG_CREATION';
+
+export const REQUEST_REMOVE_TAG = 'REQUEST_REMOVE_TAG';
+export const REQUEST_REMOVE_TAG_SUCCESS = 'REQUEST_REMOVE_TAG_SUCCESS';
+export const REQUEST_REMOVE_TAG_FAILURE = 'REQUEST_REMOVE_TAG_FAILURE';
