+++ client/app/reader/DocumentListHeader.jsx
@@ -58,9 +58,9 @@ DocumentListHeader.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  numberOfDocuments: state.readerReducer.ui.filteredDocIds ?
-    state.readerReducer.ui.filteredDocIds.length : _.size(state.documents),
-  docFilterCriteria: state.readerReducer.ui.docFilterCriteria,
+  numberOfDocuments: state.documentList.filteredDocIds ?
+    state.documentList.length : _.size(state.documents),
+  docFilterCriteria: state.documentList.docFilterCriteria,
   vacolsId: state.readerReducer.loadedAppealId
 });
