+++ client/app/reader/constants.js
@@ -2,45 +2,21 @@ import { CategoryIcon } from '../components/RenderFunctions';
 import React from 'react';
 
 // actions
-export const TOGGLE_DOCUMENT_CATEGORY = 'TOGGLE_DOCUMENT_CATEGORY';
-export const TOGGLE_DOCUMENT_CATEGORY_FAIL = 'TOGGLE_DOCUMENT_CATEGORY_FAIL';
-export const RECEIVE_DOCUMENTS = 'RECEIVE_DOCUMENTS';
-export const RECEIVE_MANIFESTS = 'RECEIVE_MANIFESTS';
 export const RECEIVE_ANNOTATIONS = 'RECEIVE_ANNOTATIONS';
 export const RECEIVE_APPEAL_DETAILS = 'RECEIVE_APPEAL_DETAILS';
 export const RECEIVE_APPEAL_DETAILS_FAILURE = 'RECEIVE_APPEAL_DETAILS_FAILURE';
 export const REQUEST_INITIAL_DATA_FAILURE = 'INITIAL_DATA_LOADING_FAIL';
 export const REQUEST_INITIAL_CASE_FAILURE = 'INITIAL_CASE_LOADING_FAIL';
-export const TOGGLE_FILTER_DROPDOWN = 'TOGGLE_FILTER_DROPDOWN';
-export const SET_CATEGORY_FILTER = 'SET_CATEGORY_FILTER';
-export const SET_TAG_FILTER = 'SET_TAG_FILTER';
-export const CLEAR_CATEGORY_FILTER = 'CLEAR_CATEGORY_FILTER';
-export const CLEAR_TAG_FILTER = 'CLEAR_TAG_FILTER';
-export const ADD_NEW_TAG = 'ADD_NEW_TAG';
-export const REMOVE_TAG = 'REMOVE_TAG';
-export const REQUEST_NEW_TAG_CREATION_SUCCESS = 'REQUEST_NEW_TAG_CREATION_SUCCESS';
-export const REQUEST_NEW_TAG_CREATION_FAILURE = 'REQUEST_NEW_TAG_CREATION_FAILURE';
-export const REQUEST_NEW_TAG_CREATION = 'REQUEST_NEW_TAG_CREATION';
-export const REQUEST_REMOVE_TAG = 'REQUEST_REMOVE_TAG';
-export const REQUEST_REMOVE_TAG_SUCCESS = 'REQUEST_REMOVE_TAG_SUCCESS';
-export const REQUEST_REMOVE_TAG_FAILURE = 'REQUEST_REMOVE_TAG_FAILURE';
-export const SELECT_CURRENT_VIEWER_PDF = 'SELECT_CURRENT_VIEWER_PDF';
+
 export const SHOW_NEXT_PDF = 'SHOW_NEXT_PDF';
 export const SHOW_PREV_PDF = 'SHOW_PREV_PDF';
 export const SCROLL_TO_COMMENT = 'SCROLL_TO_COMMENT';
-export const TOGGLE_COMMENT_LIST = 'TOGGLE_COMMENT_LIST';
 export const TOGGLE_PDF_SIDEBAR = 'TOGGLE_PDF_SIDEBAR';
 export const TOGGLE_SEARCH_BAR = 'TOGGLE_SEARCH_BAR';
 export const SHOW_SEARCH_BAR = 'SHOW_SEARCH_BAR';
 export const HIDE_SEARCH_BAR = 'HIDE_SEARCH_BAR';
-export const LAST_READ_DOCUMENT = 'LAST_READ_DOCUMENT';
 export const SCROLL_TO_SIDEBAR_COMMENT = 'SCROLL_TO_SIDEBAR_COMMENT';
 export const COLLECT_ALL_TAGS_FOR_OPTIONS = 'COLLECT_ALL_TAGS_FOR_OPTIONS';
-export const SET_SORT = 'SET_SORT';
-export const SET_SEARCH = 'SET_SEARCH';
-export const CLEAR_ALL_FILTERS = 'CLEAR_ALL_FILTERS';
-export const CLEAR_ALL_SEARCH = 'CLEAR_ALL_SEARCH';
-export const SET_DOC_LIST_SCROLL_POSITION = 'SET_DOC_LIST_SCROLL_POSITION';
 export const SET_DOC_SCROLL_POSITION = 'SET_DOC_SCROLL_POSITION';
 export const JUMP_TO_PAGE = 'JUMP_TO_PAGE';
 export const RESET_JUMP_TO_PAGE = 'RESET_JUMP_TO_PAGE';
@@ -48,7 +24,6 @@ export const RESET_JUMP_TO_PAGE = 'RESET_JUMP_TO_PAGE';
 export const SET_PAGE_COORD_BOUNDS = 'SET_PAGE_COORD_BOUNDS';
 export const SET_UP_PAGE_DIMENSIONS = 'SET_UP_PAGE_DIMENSIONS';
 export const SET_PDF_DOCUMENT = 'SET_PDF_DOCUMENT';
-export const ROTATE_PDF_DOCUMENT = 'ROTATE_PDF_DOCUMENT';
 export const CLEAR_PDF_DOCUMENT = 'CLEAR_PDF_DOCUMENT';
 export const GET_DOCUMENT_TEXT = 'GET_DCOUMENT_TEXT';
 export const ZERO_SEARCH_INDEX = 'ZERO_SEARCH_INDEX';
@@ -58,23 +33,14 @@ export const SET_SEARCH_INDEX_TO_HIGHLIGHT = 'SET_SEARCH_INDEX_TO_HIGHLIGHT';
 export const SET_DOCUMENT_LOAD_ERROR = 'SET_DOCUMENT_LOAD_ERROR';
 export const CLEAR_DOCUMENT_LOAD_ERROR = 'CLEAR_DOCUMENT_LOAD_ERROR';
 export const SET_LOADED_APPEAL_ID = 'SET_LOADED_APPEAL_ID';
-export const UPDATE_FILTERED_RESULTS = 'UPDATE_FILTERED_RESULTS';
-export const ASSIGN_DOCUMENTS = 'ASSIGN_DOCUMENTS';
 
 export const SET_OPENED_ACCORDION_SECTIONS = 'SET_OPENED_ACCORDION_SECTIONS';
-export const SET_VIEWING_DOCUMENTS_OR_COMMENTS = 'SET_VIEWING_DOCUMENTS_OR_COMMENTS';
 
 export const COMMENT_ACCORDION_KEY = 'Comments';
-
-export const DOCUMENTS_OR_COMMENTS_ENUM = {
-  DOCUMENTS: 'documents',
-  COMMENTS: 'comments'
-};
 
 // If we used CSS in JS, we wouldn't have to keep this value in sync with the CSS in a brittle way.
 export const ANNOTATION_ICON_SIDE_LENGTH = 40;
 
-export const ROTATION_INCREMENTS = 90;
 export const COMPLETE_ROTATION = 360;
 
 // Arrange the directions such that each direction + 1 modulo 4 rotates clockwise.
