+++ client/app/reader/index.jsx
@@ -10,6 +10,7 @@ import { getReduxAnalyticsMiddleware } from '../util/getReduxAnalyticsMiddleware
 import { reducer as searchReducer, reduxSearch } from 'redux-search';
 import { annotationLayerReducer } from './AnnotationLayer/AnnotationLayerReducer';
 import documentsReducer from './DocumentList/DocumentsReducer';
+import documentListReducer from './DocumentList/DocumentListReducer';
 
 // eslint-disable-next-line no-underscore-dangle
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
@@ -19,7 +20,8 @@ const store = createStore(
     readerReducer,
     search: searchReducer,
     documents: documentsReducer,
-    annotationLayer: annotationLayerReducer
+    annotationLayer: annotationLayerReducer,
+    documentList: documentListReducer
   }),
   composeEnhancers(
     applyMiddleware(thunk, perfLogger, getReduxAnalyticsMiddleware()),
